Change behavior of after to ignore cache if handling an error
For two reasons:
Because somewhere in the Restify execution chain, the rs.writeHead() method is invoked, thus locking the headers. If the after method were to fire, writing the X-Cache header will fail.
It seems unlikely that writing an error to cache is useful, especially if some middleware already passed an error into next or may have already wrote exception data to the response.